requirements.txt
alembic
boto3
fastapi
pq
psycopg2
pydantic
smart_open
toposort
typeguard
ulid-py
uvicorn[standard]
